requirements.txt
# ICD-10 诊断标准化RAG系统

# ========================
# AI和机器学习核心组件
# ========================
# 向量化和语言模型
sentence-transformers>=4.1.0
torch>=2.2.0
numpy>=1.25.2

# sentence-transformers的必需依赖
scikit-learn>=1.3.0
scipy>=1.10.0
joblib>=1.3.0
pillow>=10.0.0

# torch的必需依赖
sympy>=1.12.0
networkx>=3.0.0

# transformers的必需依赖
regex>=2023.0.0

# ========================
# Web框架和API
# ========================
# FastAPI Web框架
fastapi==0.104.1
uvicorn==0.24.0

# ========================
# 向量数据库
# ========================
# Milvus向量数据库
pymilvus==2.5.10

# ========================
# 数据处理
# ========================
# 数据分析
pandas==2.1.3

# 数据验证
pydantic==2.5.0

# 进度条
tqdm==4.66.1

# ========================
# LLM API客户端
# ========================
# OpenAI API
openai>=1.55.3

# ========================
# 配置和环境管理
# ========================
# 环境变量
python-dotenv==1.1.0

# ========================
# 日志管理
# ========================
# 日志
loguru==0.7.2

# ========================
# UI界面
# ========================
# Gradio Web UI
gradio>=4.44.0
